changes dealing with circular import
